requirements.txt
langchain
python-dotenv
gunicorn
uvicorn
openai
spotipy
tenacity
requests
langchain_community